+++ R/metabin.R
@@ -181,7 +181,7 @@
 #'   should be kept in meta object.
 #' @param warn A logical indicating whether warnings should be printed
 #'   (e.g., if \code{incr} is added to studies with zero cell
-#'   frequencies).
+#'   frequencies or if estimation problems exist in fitting a GLMM).
 #' @param warn.deprecated A logical indicating whether warnings should
 #'   be printed if deprecated arguments are used.
 #' @param control An optional list to control the iterative process to
@@ -271,7 +271,7 @@
 #' R package \bold{brglm2} must be available to fit a one-stage logistic
 #' regression model with penalised likelihood (Evrenoglou et al., 2022).
 #' The estimation of the summary odds ratio relies on the maximisation of the 
-#' likelihood function, penalized using a Firth-type correction. This
+#' likelihood function, penalised using a Firth-type correction. This
 #' penalisation aims to reduce bias in cases with rare events and a small
 #' number of available studies. However, this method is not restricted 
 #' to only such cases and can be applied more generally to binary data. Note,
@@ -1702,7 +1702,8 @@ metabin <- function(event.e, n.e, event.c, n.c, studlab,
                   call. = FALSE)
       }
     else if (is.glmm | is.lrp) {
-      if ((sparse | addincr) & warn)
+      if (sparse & warn &
+          ((!missing.incr & any(incr != 0)) | allincr | addincr))
         warning("Note, for method = \"", method,
                 "\", continuity correction ",
                 "only used to calculate individual study results.",
@@ -1824,7 +1825,8 @@ metabin <- function(event.e, n.e, event.c, n.c, studlab,
               method.random.ci = method.random.ci,
               level = level.ma,
               control = list(control),
-              use.random = use.random)
+              use.random = use.random,
+              warn = warn)
     ##
     TE.common   <- as.numeric(res.glmm$glmm.common$b)
     seTE.common <- as.numeric(res.glmm$glmm.common$se)
@@ -1833,7 +1835,8 @@ metabin <- function(event.e, n.e, event.c, n.c, studlab,
   }
   else if (is.lrp) {
     fit.lrp <- runLRP(event.e[!exclude], n1 = n.e[!exclude],
-                      event2 = event.c[!exclude], n2 = n.c[!exclude])
+                      event2 = event.c[!exclude], n2 = n.c[!exclude],
+                      ...)
     #
     TE.common   <- fit.lrp$TE.common
     seTE.common <- fit.lrp$seTE.common
@@ -1863,7 +1866,8 @@ metabin <- function(event.e, n.e, event.c, n.c, studlab,
                overall.hetstat = overall.hetstat,
                prediction = prediction,
                ##
-               method.tau = method.tau, method.tau.ci = method.tau.ci,
+               method.tau = if (is.glmm | is.lrp) "DL" else method.tau,
+               method.tau.ci = if (is.glmm | is.lrp) "" else method.tau.ci,
                level.hetstat = level.hetstat,
                tau.preset = tau.preset,
                TE.tau = if (Q.Cochrane) TE.common else TE.tau,
@@ -1972,6 +1976,7 @@ metabin <- function(event.e, n.e, event.c, n.c, studlab,
   }
   #
   if (is.glmm) {
+    res$method.tau <- method.tau
     res <- addGLMM(res, res.glmm, method.I2)
     res$model.glmm <- model.glmm
     ##
@@ -1998,7 +2003,8 @@ metabin <- function(event.e, n.e, event.c, n.c, studlab,
                     else
                       NULL,
                   control = list(control),
-                  use.random = use.random)$glmm.random[[1]],
+                  use.random = use.random,
+                  warn = warn)$glmm.random[[1]],
           method.I2
         )
     }
@@ -2081,7 +2087,7 @@ metabin <- function(event.e, n.e, event.c, n.c, studlab,
   if (res$method.random == "MH")
     res$method.random <- "Inverse"
   #
-  # Do not return tau^2 and tau for penalized logistic regression
+  # Do not return tau^2 and tau for penalised logistic regression
   #
   if (is.lrp) {
     res$method.random <- "LRP"